Fix FloatColumn#getDataType

+++ tsfile/src/main/java/org/apache/iotdb/tsfile/read/common/block/column/FloatColumn.java
@@ -69,7 +69,7 @@ public class FloatColumn implements Column {
 
   @Override
   public TSDataType getDataType() {
-    return TSDataType.DOUBLE;
+    return TSDataType.FLOAT;
   }
 
   @Override
